+++ ocd-3148.sql
@@ -1,5 +1,0 @@
-ALTER TABLE openchpl.certification_criterion
-DROP COLUMN IF EXISTS removed;
-
-ALTER TABLE openchpl.certification_criterion
-ADD COLUMN removed boolean NOT NULL DEFAULT false;