+++ v-next.sql
@@ -1,7 +1,3 @@
-\i ocd-3188.sql
-\i ocd-3135.sql
-\i ocd-3148.sql
-
 \i dev/openchpl_soft-delete.sql
 \i dev/openchpl_views.sql
 \i dev/openchpl_grant-all.sql